detail -> reflect, move type_tratis to base dir
