+++ include/elastic/types/varint.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "../detail/concepts.hpp"
+#include "../concepts.hpp"
 
 namespace elastic
 {
