+++ test/test.vcxproj
@@ -145,17 +145,16 @@
     <ClInclude Include="..\include\elastic\binary_iserialize.hpp" />
     <ClInclude Include="..\include\elastic\binary_oarchive.hpp" />
     <ClInclude Include="..\include\elastic\binary_oserialize.hpp" />
-    <ClInclude Include="..\include\elastic\detail\concepts.hpp" />
-    <ClInclude Include="..\include\elastic\detail\filed_count.hpp" />
-    <ClInclude Include="..\include\elastic\detail\generate.hpp" />
-    <ClInclude Include="..\include\elastic\detail\nocopyable.hpp" />
-    <ClInclude Include="..\include\elastic\detail\size_t_.hpp" />
-    <ClInclude Include="..\include\elastic\detail\type_traits.hpp" />
+    <ClInclude Include="..\include\elastic\concepts.hpp" />
     <ClInclude Include="..\include\elastic\interface_archive.hpp" />
     <ClInclude Include="..\include\elastic\property\fixed.hpp" />
     <ClInclude Include="..\include\elastic\property\optional.hpp" />
     <ClInclude Include="..\include\elastic\property\repeated.hpp" />
     <ClInclude Include="..\include\elastic\reflect.hpp" />
+    <ClInclude Include="..\include\elastic\reflect\filed_count.hpp" />
+    <ClInclude Include="..\include\elastic\reflect\generate.hpp" />
+    <ClInclude Include="..\include\elastic\reflect\nocopyable.hpp" />
+    <ClInclude Include="..\include\elastic\reflect\size_t_.hpp" />
     <ClInclude Include="..\include\elastic\tuple_size.hpp" />
     <ClInclude Include="..\include\elastic\types\bytes.hpp" />
     <ClInclude Include="..\include\elastic\types\float.hpp" />
@@ -163,6 +162,7 @@
     <ClInclude Include="..\include\elastic\types\sequence.hpp" />
     <ClInclude Include="..\include\elastic\types\string.hpp" />
     <ClInclude Include="..\include\elastic\types\varint.hpp" />
+    <ClInclude Include="..\include\elastic\type_traits.hpp" />
     <ClInclude Include="attr.h" />
     <ClInclude Include="exception.h" />
     <ClInclude Include="generate.h" />
